refactor(helper): move `helper` dir into `common` dir
